Doing "ReactApp - Tic Tac Toe": - Moved profile constant to app ones
